Split type checker into modules
